Refactor code and add weekend/weekday calculation functions
The code was extensively refactored with changes mainly focused on improving the predictions handling. The refactoring has also led to the introduction of two new functions (get_week_nights and get_weekend_nights) that calculate the number of weeknights and weekend nights between two dates respectively. This provides better, more detailed user input which in turn results in more accurate predictions. The input collection from the user was overhauled and replaced with a form that verifies the input before submitting. In addition, use of Streamlit's session state to store the predictive model was implemented to avoid reloading the model for every new prediction.